+++ server/controllers/mypage/withdrawal.js
@@ -1,55 +1,41 @@
-// const { isAuthorized } = require('../tokenFunctions');
+const { isAuthorized } = require('../tokenFunctions');
 const { user } = require('../../models');
 
 // DELETE http://localhost:80/withdrawal
 module.exports = async (req, res) => {
   try {
-    // const accessTokenData = isAuthorized(req);
+    const accessTokenData = isAuthorized(req);
 
-    // if (!accessTokenData) {
-    //   return res.status(404).send({ message: 'You\'re not logged in.' });
-    // } else {
-    //   await User.destroy({
-    //     where: { id: accessTokenData.id },
-    //   });
-    // }
-
-    // ===============================================================
-    //                       JUST FOR TEST PURPOSES
-    // ================================================================
-    
-    console.log('🥺🥺🥺🥺🥺🥺🥺');
-    console.log(req.headers.authorization);
-
-    let userInfo = await user.findOne({
-      where: {
-        id: req.headers.authorization
-      }
-    });
-
-    // ================================================================
-
-    if (userInfo) {
-      await user.destroy({
+    if (!accessTokenData) {
+      return res.status(401).send({ message: 'You\'re not logged in.' });
+    } else {
+      const userInfo = await user.findOne({
         where: {
-          id: userInfo.id
+          id: accessTokenData.id
         }
       });
 
-      // res.setHeader('authorization', '');
-      // res.cookie('refreshToken', '');
+      if (userInfo) {
+        await user.destroy({
+          where: {
+            id: accessTokenData.id
+          }
+        });
 
-      res.status(200).json({
-        message: 'Successfully withdrawn'
-      });
-    } else {
-      res.status(400).json({
-        message: 'Invalid access token'
-      });
+        res.setHeader('authorization', '');
+
+        res.status(200).json({
+          message: 'Successfully withdrawn'
+        });
+      } else {
+        res.status(400).json({
+          message: 'error'
+        });
+      }
     }
   } catch {
     res.status(400).json({
-      message: 'Invalid access token'
+      message: 'error'
     });
   }
 };
